requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.30.0
# Upgrade pydantic to satisfy autogen-core (>=2.10.0) & mcp (>=2.8.0)
pydantic>=2.10.0,<3.0.0
langchain~=0.2.10
langgraph~=0.1.8
openai~=1.37.0
httpx~=0.27.0
python-dotenv~=1.0.1
pytest~=8.3.0
langchain-openai>=0.1.7
